requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
imageio==2.9.0
importlib-metadata==2.0.0
joblib==0.17.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
labelme==3.16.7
Markdown==3.3
matplotlib==3.3.2
networkx==2.5
nibabel==3.1.1
numpy==1.19.2
opencv-python==4.4.0.44
packaging==20.4
pandas==1.1.3
Pillow==7.2.0
protobuf==3.13.0
pycococreatortools @ git+git://github.com/waspinator/pycococreator.git@fba8f4098f3c7aaa05fe119dc93bbe4063afdab8
pycocotools==2.0.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
QtPy==1.9.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
SimpleITK==2.0.0
six==1.15.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
